Fix primary key issue
It's bad practice to change the primary key, which we were allowing with
the `name` field. Adding an `ID` primary key and switching to a UNIQUE
constraint on `name` gives us the same functionality without violating
any db principles.

Note: This can't be migrated automatically, you'll have to delete any
existing database (or migrate it manually).